Split code (#29)
* split search types

* lint

* lint

* readme changes

* documentation
