+++ examples/keras_example.py
@@ -8,7 +8,7 @@ from keras.layers import Dense, Activation, Dropout
 from keras.models import Sequential
 from keras.utils import np_utils
 from keras.wrappers.scikit_learn import KerasClassifier
-from tune_sklearn.tune_search import TuneGridSearchCV
+from tune_sklearn.tune_gridsearch import TuneGridSearchCV
 
 nb_classes = 10
 (X_train, y_train), (X_test, y_test) = mnist.load_data()
