+++ tests/test_gridsearch.py
@@ -1,4 +1,5 @@
-from tune_sklearn import tune_search as tcv
+from tune_sklearn.tune_gridsearch import TuneGridSearchCV
+from tune_sklearn.tune_search import TuneSearchCV
 import numpy as np
 from numpy.testing import (
     assert_almost_equal,
@@ -82,7 +83,7 @@ class GridSearchTest(unittest.TestCase):
     def test_grid_search(self):
         # Test that the best estimator contains the right value for foo_param
         clf = MockClassifier()
-        grid_search = tcv.TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]}, cv=3)
+        grid_search = TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]}, cv=3)
         # make sure it selects the smallest parameter in case of ties
         grid_search.fit(X, y)
         self.assertEqual(grid_search.best_estimator_.foo_param, 2)
@@ -112,11 +113,11 @@ class GridSearchTest(unittest.TestCase):
         # multiple `SVC` instances in parallel using threads sometimes results
         # in wrong results. This only happens with threads, not processes/sync.
         # For now, we'll fit using the sync scheduler.
-        grid_search = tcv.TuneGridSearchCV(
+        grid_search = TuneGridSearchCV(
             clf, {"C": Cs}, scoring="accuracy", scheduler=MedianStoppingRule())
         grid_search.fit(X, y)
 
-        grid_search_no_score = tcv.TuneGridSearchCV(
+        grid_search_no_score = TuneGridSearchCV(
             clf_no_score, {"C": Cs},
             scoring="accuracy",
             scheduler=MedianStoppingRule())
@@ -131,16 +132,16 @@ class GridSearchTest(unittest.TestCase):
             grid_search.score(X, y), grid_search_no_score.score(X, y))
 
         # giving no scoring function raises an error
-        grid_search_no_score = tcv.TuneGridSearchCV(clf_no_score, {"C": Cs})
+        grid_search_no_score = TuneGridSearchCV(clf_no_score, {"C": Cs})
         with self.assertRaises(TypeError) as exc:
             grid_search_no_score.fit([[1]])
 
         self.assertTrue("no scoring" in str(exc.exception))
 
-    @parameterized.expand([("grid", tcv.TuneGridSearchCV, {}),
-                           ("random", tcv.TuneSearchCV, {
-                               "n_iter": 1
-                           })])
+    @parameterized.expand([("grid", TuneGridSearchCV, {}), ("random",
+                                                            TuneSearchCV, {
+                                                                "n_iter": 1
+                                                            })])
     def test_hyperparameter_searcher_with_fit_params(self, name, cls, kwargs):
         X = np.arange(100).reshape(10, 10)
         y = np.array([0] * 5 + [1] * 5)
@@ -163,14 +164,12 @@ class GridSearchTest(unittest.TestCase):
         clf = LinearSVC(random_state=0)
         grid = {"C": [0.1]}
 
-        search_no_scoring = tcv.TuneGridSearchCV(
-            clf, grid, scoring=None).fit(X, y)
-        search_accuracy = tcv.TuneGridSearchCV(
+        search_no_scoring = TuneGridSearchCV(clf, grid, scoring=None).fit(X, y)
+        search_accuracy = TuneGridSearchCV(
             clf, grid, scoring="accuracy").fit(X, y)
-        search_no_score_method_auc = tcv.TuneGridSearchCV(
+        search_no_score_method_auc = TuneGridSearchCV(
             LinearSVCNoScore(), grid, scoring="roc_auc").fit(X, y)
-        search_auc = tcv.TuneGridSearchCV(
-            clf, grid, scoring="roc_auc").fit(X, y)
+        search_auc = TuneGridSearchCV(clf, grid, scoring="roc_auc").fit(X, y)
 
         # Check warning only occurs in situation where behavior changed:
         # estimator requires score method to compete with scoring parameter
@@ -206,7 +205,7 @@ class GridSearchTest(unittest.TestCase):
             GroupShuffleSplit(n_splits=3),
         ]
         for cv in group_cvs:
-            gs = tcv.TuneGridSearchCV(clf, grid, cv=cv)
+            gs = TuneGridSearchCV(clf, grid, cv=cv)
 
             with self.assertRaises((ValueError, TuneError)) as exc:
                 gs.fit(X, y)
@@ -221,7 +220,7 @@ class GridSearchTest(unittest.TestCase):
             StratifiedShuffleSplit(n_splits=3)
         ]
         for cv in non_group_cvs:
-            gs = tcv.TuneGridSearchCV(clf, grid, cv=cv)
+            gs = TuneGridSearchCV(clf, grid, cv=cv)
             # Should not raise an error
             gs.fit(X, y)
 
@@ -233,24 +232,22 @@ class GridSearchTest(unittest.TestCase):
         y = np.array([0] * 5 + [1] * 5)
         Cs = [0.1, 1, 10]
 
-        grid_search = tcv.TuneGridSearchCV(
-            LinearSVC(random_state=0), {"C": Cs})
+        grid_search = TuneGridSearchCV(LinearSVC(random_state=0), {"C": Cs})
         grid_search.fit(X, y)
         assert_array_equal(grid_search.best_estimator_.classes_,
                            grid_search.classes_)
 
         # Test that regressors do not have a classes_ attribute
-        grid_search = tcv.TuneGridSearchCV(Ridge(), {"alpha": [1.0, 2.0]})
+        grid_search = TuneGridSearchCV(Ridge(), {"alpha": [1.0, 2.0]})
         grid_search.fit(X, y)
         self.assertFalse(hasattr(grid_search, "classes_"))
 
         # Test that the grid searcher has no classes_ attribute before it's fit
-        grid_search = tcv.TuneGridSearchCV(
-            LinearSVC(random_state=0), {"C": Cs})
+        grid_search = TuneGridSearchCV(LinearSVC(random_state=0), {"C": Cs})
         self.assertFalse(hasattr(grid_search, "classes_"))
 
         # Test that the grid searcher has no classes_ attribute without a refit
-        grid_search = tcv.TuneGridSearchCV(
+        grid_search = TuneGridSearchCV(
             LinearSVC(random_state=0), {"C": Cs}, refit=False)
         grid_search.fit(X, y)
         self.assertFalse(hasattr(grid_search, "classes_"))
@@ -260,12 +257,11 @@ class GridSearchTest(unittest.TestCase):
         # Non-regression test: grid_scores_ wouldn't be set by
         # dcv.GridSearchCV.
         clf = MockClassifier()
-        grid_search = tcv.TuneGridSearchCV(clf, {"foo_param": [1]}, cv=3)
+        grid_search = TuneGridSearchCV(clf, {"foo_param": [1]}, cv=3)
         grid_search.fit(X, y)
         self.assertTrue(hasattr(grid_search, "cv_results_"))
 
-        random_search = tcv.TuneSearchCV(
-            clf, {"foo_param": [0]}, n_iter=1, cv=3)
+        random_search = TuneSearchCV(clf, {"foo_param": [0]}, n_iter=1, cv=3)
         random_search.fit(X, y)
         self.assertTrue(hasattr(random_search, "cv_results_"))
 
@@ -273,7 +269,7 @@ class GridSearchTest(unittest.TestCase):
         # Test that GSCV can be used for model selection alone without
         # refitting
         clf = MockClassifier()
-        grid_search = tcv.TuneGridSearchCV(
+        grid_search = TuneGridSearchCV(
             clf, {"foo_param": [1, 2, 3]}, refit=False, cv=3)
         grid_search.fit(X, y)
         self.assertFalse(hasattr(grid_search, "best_estimator_"))
@@ -302,7 +298,7 @@ class GridSearchTest(unittest.TestCase):
             n_samples=200, n_features=100, random_state=0)
 
         clf = LinearSVC()
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
         with self.assertRaises(TuneError):
             cv.fit(X_[:180], y_)
 
@@ -312,7 +308,7 @@ class GridSearchTest(unittest.TestCase):
         param_dict = {"C": [1.0], "kernel": ["rbf"], "gamma": [0.1]}
 
         clf = SVC()
-        cv = tcv.TuneGridSearchCV(clf, param_dict)
+        cv = TuneGridSearchCV(clf, param_dict)
         cv.fit(X_, y_)
 
         clf = SVC(C=1.0, kernel="rbf", gamma=0.1)
@@ -325,7 +321,7 @@ class GridSearchTest(unittest.TestCase):
         clf = SVC()
 
         with self.assertRaises(ValueError) as exc:
-            tcv.TuneGridSearchCV(clf, param_dict)
+            TuneGridSearchCV(clf, param_dict)
         self.assertTrue(("Parameter grid for parameter (C) needs to"
                          " be a list or numpy array") in str(exc.exception))
 
@@ -333,7 +329,7 @@ class GridSearchTest(unittest.TestCase):
         clf = SVC()
 
         with self.assertRaises(ValueError) as exc:
-            tcv.TuneGridSearchCV(clf, param_dict)
+            TuneGridSearchCV(clf, param_dict)
         self.assertTrue((
             "Parameter values for parameter (C) need to be a non-empty "
             "sequence.") in str(exc.exception))
@@ -342,14 +338,14 @@ class GridSearchTest(unittest.TestCase):
         clf = SVC()
 
         with self.assertRaises(ValueError) as exc:
-            tcv.TuneGridSearchCV(clf, param_dict)
+            TuneGridSearchCV(clf, param_dict)
         self.assertTrue(("Parameter grid for parameter (C) needs to"
                          " be a list or numpy array") in str(exc.exception))
 
         param_dict = {"C": np.ones(6).reshape(3, 2)}
         clf = SVC()
         with self.assertRaises(ValueError):
-            tcv.TuneGridSearchCV(clf, param_dict)
+            TuneGridSearchCV(clf, param_dict)
 
     def test_grid_search_sparse(self):
         # Test that grid search works with both dense and sparse matrices
@@ -357,14 +353,14 @@ class GridSearchTest(unittest.TestCase):
             n_samples=200, n_features=100, random_state=0)
 
         clf = LinearSVC()
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
         cv.fit(X_[:180], y_[:180])
         y_pred = cv.predict(X_[180:])
         C = cv.best_estimator_.C
 
         X_ = sp.csr_matrix(X_)
         clf = LinearSVC()
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
         cv.fit(X_[:180].tocoo(), y_[:180])
         y_pred2 = cv.predict(X_[180:])
         C2 = cv.best_estimator_.C
@@ -377,14 +373,14 @@ class GridSearchTest(unittest.TestCase):
             n_samples=200, n_features=100, random_state=0)
 
         clf = LinearSVC()
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]}, scoring="f1")
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]}, scoring="f1")
         cv.fit(X_[:180], y_[:180])
         y_pred = cv.predict(X_[180:])
         C = cv.best_estimator_.C
 
         X_ = sp.csr_matrix(X_)
         clf = LinearSVC()
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]}, scoring="f1")
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]}, scoring="f1")
         cv.fit(X_[:180], y_[:180])
         y_pred2 = cv.predict(X_[180:])
         C2 = cv.best_estimator_.C
@@ -397,7 +393,7 @@ class GridSearchTest(unittest.TestCase):
             return -f1_score(y_true_, y_pred_)
 
         F1Loss = make_scorer(f1_loss, greater_is_better=False)
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]}, scoring=F1Loss)
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]}, scoring=F1Loss)
         cv.fit(X_[:180], y_[:180])
         y_pred3 = cv.predict(X_[180:])
         C3 = cv.best_estimator_.C
@@ -416,7 +412,7 @@ class GridSearchTest(unittest.TestCase):
         y_train = y_[:180]
 
         clf = SVC(kernel="precomputed")
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
         cv.fit(K_train, y_train)
 
         self.assertTrue(cv.best_score_ >= 0)
@@ -440,7 +436,7 @@ class GridSearchTest(unittest.TestCase):
         K_train = np.zeros((10, 20))
         y_train = np.ones((10, ))
         clf = SVC(kernel="precomputed")
-        cv = tcv.TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
+        cv = TuneGridSearchCV(clf, {"C": [0.1, 1.0]})
         with self.assertRaises(TuneError):
             cv.fit(K_train, y_train)
 
@@ -451,7 +447,7 @@ class GridSearchTest(unittest.TestCase):
         X = np.arange(100).reshape(10, 10)
         y = np.array([0] * 5 + [1] * 5)
 
-        clf = tcv.TuneGridSearchCV(
+        clf = TuneGridSearchCV(
             BrokenClassifier(), {"parameter": [0, 1]},
             scoring="accuracy",
             refit=True)
@@ -465,7 +461,7 @@ class GridSearchTest(unittest.TestCase):
             check_X=lambda x: x.shape[1:] == (5, 3, 2),
             check_y=lambda x: x.shape[1:] == (7, 11),
         )
-        grid_search = tcv.TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]}, cv=3)
+        grid_search = TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]}, cv=3)
         grid_search.fit(X_4d, y_3d).score(X, y)
         self.assertTrue(hasattr(grid_search, "cv_results_"))
 
@@ -476,8 +472,7 @@ class GridSearchTest(unittest.TestCase):
 
         clf = CheckingClassifier(check_X=lambda x: isinstance(x, list))
         cv = KFold(n_splits=3)
-        grid_search = tcv.TuneGridSearchCV(
-            clf, {"foo_param": [1, 2, 3]}, cv=cv)
+        grid_search = TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]}, cv=cv)
         grid_search.fit(X.tolist(), y).score(X, y)
         self.assertTrue(hasattr(grid_search, "cv_results_"))
 
@@ -488,8 +483,7 @@ class GridSearchTest(unittest.TestCase):
 
         clf = CheckingClassifier(check_y=lambda x: isinstance(x, list))
         cv = KFold(n_splits=3)
-        grid_search = tcv.TuneGridSearchCV(
-            clf, {"foo_param": [1, 2, 3]}, cv=cv)
+        grid_search = TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]}, cv=cv)
         grid_search.fit(X, y.tolist()).score(X, y)
         self.assertTrue(hasattr(grid_search, "cv_results_"))
 
@@ -515,7 +509,7 @@ class GridSearchTest(unittest.TestCase):
                 check_y=lambda x: isinstance(x, TargetType),
             )
 
-            grid_search = tcv.TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]})
+            grid_search = TuneGridSearchCV(clf, {"foo_param": [1, 2, 3]})
             grid_search.fit(X_df, y_ser).score(X_df, y_ser)
             grid_search.predict(X_df)
             self.assertTrue(hasattr(grid_search, "cv_results_"))
@@ -524,7 +518,7 @@ class GridSearchTest(unittest.TestCase):
         # test grid-search with unsupervised estimator
         X, y = make_blobs(random_state=0)
         km = KMeans(random_state=0)
-        grid_search = tcv.TuneGridSearchCV(
+        grid_search = TuneGridSearchCV(
             km,
             param_grid=dict(n_clusters=[2, 3, 4]),
             scoring="adjusted_rand_score")
@@ -533,7 +527,7 @@ class GridSearchTest(unittest.TestCase):
         self.assertEqual(grid_search.best_params_["n_clusters"], 3)
 
         # Now without a score, and without y
-        grid_search = tcv.TuneGridSearchCV(
+        grid_search = TuneGridSearchCV(
             km, param_grid=dict(n_clusters=[2, 3, 4]))
         grid_search.fit(X)
         self.assertEqual(grid_search.best_params_["n_clusters"], 4)
@@ -546,7 +540,7 @@ class GridSearchTest(unittest.TestCase):
 
         X, _ = make_blobs(
             cluster_std=0.1, random_state=1, centers=[[0, 1], [1, 0], [0, 0]])
-        search = tcv.TuneGridSearchCV(
+        search = TuneGridSearchCV(
             KernelDensity(),
             param_grid=dict(bandwidth=[0.01, 0.1, 1]),
             scoring=custom_scoring,
@@ -576,7 +570,7 @@ class GridSearchTest(unittest.TestCase):
             "C": [1, 10, 100, 1000]
         }
 
-        tune_search = tcv.TuneGridSearchCV(
+        tune_search = TuneGridSearchCV(
             SVC(),
             tuned_parameters,
             scheduler="MedianStoppingRule",
@@ -602,8 +596,7 @@ class GridSearchTest(unittest.TestCase):
         # create and fit a ridge regression model, testing each alpha
         model = linear_model.Ridge()
 
-        tune_search = tcv.TuneGridSearchCV(model, param_grid,
-                                           "MedianStoppingRule")
+        tune_search = TuneGridSearchCV(model, param_grid, "MedianStoppingRule")
         tune_search.fit(X_train, y_train)
 
         pred = tune_search.predict(X_test)
